backend/requirements.txt
pinecone-client
openai
psycopg2-binary
redis
boto3
requests
